Fixed signed vs. unsigned comparison warning.

+++ src/ui/gtk/search_common.c
@@ -855,9 +855,9 @@ void search_matched(search_t *sch, results_set_t *rs)
         if (!(flt_result->props[FILTER_PROP_DISPLAY].state == 
                 FILTER_PROP_STATE_DONT &&
             flt_result->props[FILTER_PROP_DISPLAY].user_data == 0) &&
-			results_kept++ >= 0 &&	/* Count as kept even if max results */
-            sch->items < search_max_results)
-        {
+			(int) results_kept++ >= 0 && /* Count as kept even if max results */
+            sch->items < search_max_results
+		) {
             GdkColor *fg_color = NULL;
             gboolean mark;
